+++ assets/javascripts/kitten/components/information/alert/test.js
@@ -7,16 +7,16 @@ describe('<Alert />', () => {
     const component = renderer.create(<Alert>Alert !</Alert>).toJSON()
     expect(component).toMatchSnapshot()
   })
-  it('should matches with error snapshot', () => {
-    const component = renderer.create(<Alert error>Alert !</Alert>).toJSON()
+  it('should matches with danger snapshot', () => {
+    const component = renderer.create(<Alert status="danger">Alert !</Alert>).toJSON()
     expect(component).toMatchSnapshot()
   })
   it('should matches with success snapshot', () => {
-    const component = renderer.create(<Alert success>Alert !</Alert>).toJSON()
+    const component = renderer.create(<Alert status="success">Alert !</Alert>).toJSON()
     expect(component).toMatchSnapshot()
   })
   it('should matches with warning snapshot', () => {
-    const component = renderer.create(<Alert warning>Alert !</Alert>).toJSON()
+    const component = renderer.create(<Alert status="warning">Alert !</Alert>).toJSON()
     expect(component).toMatchSnapshot()
   })
   it('should matches with closeButton snapshot', () => {